pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sonzai"
version = "1.2.5"
description = "Python SDK for the Sonzai Mind Layer API"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Sonzai Labs", email = "engineering@sonz.ai" },
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Typing :: Typed",
]
dependencies = [
    "httpx>=0.25.0",
    "pydantic>=2.0.0",
    # Runtime dependency for classes emitted by openapi-python-client into
    # src/sonzai/_generated/. The hand-written public surface does not import
    # attrs directly, but a user poking at _generated.* will need it.
    "attrs>=23.1.0",
    # python-dateutil is emitted by the generator for datetime parsing in
    # _generated.models.*.
    "python-dateutil>=2.8.0",
]
keywords = ["sonzai", "ai", "mind-layer", "agents", "llm"]

[project.urls]
Homepage = "https://github.com/sonz-ai/sonzai-python"
Documentation = "https://github.com/sonz-ai/sonzai-python#readme"
Repository = "https://github.com/sonz-ai/sonzai-python"
Issues = "https://github.com/sonz-ai/sonzai-python/issues"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "respx>=0.21",
    "ruff>=0.4",
    "mypy>=1.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/sonzai"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
